Create separate command to open/reopen issue
