+++ app/commands/github/issue/open_for_sync_failure.rb
@@ -8,14 +8,18 @@ module Github
       def call
         return if deadlock_exception?
 
-        if missing?
-          create_issue
-        elsif closed?
-          reopen_issue
-        end
+        Github::Issue::Open.(repo, title, body)
       end
 
       private
+      def deadlock_exception?
+        exception.is_a?(ActiveRecord::Deadlocked)
+      end
+
+      def repo
+        "exercism/#{track.slug}"
+      end
+
       def title
         "🤖 Sync error for commit #{git_sha[0..5]}"
       end
@@ -33,38 +37,6 @@ module Github
 
           Please tag @exercism/maintainers-admin if you require more information.
         BODY
-      end
-
-      def create_issue
-        Exercism.octokit_client.create_issue(repo, title, body)
-      end
-
-      def reopen_issue
-        Exercism.octokit_client.reopen_issue(repo, issue.number)
-      end
-
-      def missing?
-        issue.nil?
-      end
-
-      def closed?
-        issue.state == "closed"
-      end
-
-      memoize
-      def repo
-        "exercism/#{track.slug}"
-      end
-
-      memoize
-      def issue
-        # TODO: Elevate this into exercism-config gem
-        author = "exercism-bot"
-        Exercism.octokit_client.search_issues("\"#{title}\" is:issue in:title repo:#{repo} author:#{author}")[:items]&.first
-      end
-
-      def deadlock_exception?
-        exception.is_a?(ActiveRecord::Deadlocked)
       end
     end
   end
