+++ test/commands/github/issue/open_for_sync_failure_test.rb
@@ -4,8 +4,7 @@ class Github::Issue::OpenForSyncFailureTest < ActiveSupport::TestCase
   test "opens issue when issue was not yet created" do
     head_git_sha = "2e25f799c1830b93a8ad65a2bbbb1c50f381e639"
 
-    # We raise the error and then catch it to make sure that the
-    # error has a proper backtrace
+    # Raise the error and catch it to ensure a proper backtrace
     begin
       raise StandardError, "Could not find Concept X"
     rescue StandardError => e
@@ -43,7 +42,13 @@ class Github::Issue::OpenForSyncFailureTest < ActiveSupport::TestCase
 
   test "re-opens issue when issue was closed" do
     head_git_sha = "2e25f799c1830b93a8ad65a2bbbb1c50f381e639"
-    exception = StandardError.new "Could not find Concept X"
+
+    # Raise the error and catch it to ensure a proper backtrace
+    begin
+      raise StandardError, "Could not find Concept X"
+    rescue StandardError => e
+      exception = e
+    end
 
     track = create :track, slug: 'ruby'
     track.stubs(:git_head_sha).returns(head_git_sha)
@@ -72,7 +77,13 @@ class Github::Issue::OpenForSyncFailureTest < ActiveSupport::TestCase
 
   test "does nothing when issue already open" do
     head_git_sha = "2e25f799c1830b93a8ad65a2bbbb1c50f381e639"
-    exception = StandardError.new "Could not find Concept X"
+
+    # Raise the error and catch it to ensure a proper backtrace
+    begin
+      raise StandardError, "Could not find Concept X"
+    rescue StandardError => e
+      exception = e
+    end
 
     track = create :track, slug: 'ruby'
     track.stubs(:git_head_sha).returns(head_git_sha)
